Completed Exercises for Fundamentals Part 4
Because
- It is part of the syllabus for The Odin Project

I completed the following exercises and passed all of the spec
tests: helloWorld, repeatString, reverseString, removeFromArray,
sumAll, leapYears, and tempConversion. I found removeFromArray to
be the most difficult. I know how I would have done it in Python but
it took some googling in order to figure out how to do the same
thing in javascript.